requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
phonenumbers==9.0.8
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
xlwt==1.3.0
